Update data.js
Added copy 

At the Humane Sleep Project, we are dedicated to addressing challenges faced by the unsheltered population of California. 

With 135,771 individuals affected each night, it constitutes the majority of the nation's homeless population. 

Through our outreach efforts, we seek to establish connections with those in need and offer a listening ear and essential supplies through drives and direct donations. Long-term, we look to create opportunities for sustainable housing and practices to support the unsheltered population.